requirements.txt
flask==2.0.1
gunicorn==20.1.0
gevent==21.12.0
requests==2.27.1
pyjwt==2.3.0
python-dotenv==0.19.2
psycopg2-binary==2.9.6
